debug handling negative number, change return to float

+++ 0x08-recursion/3-factorial.c
@@ -4,7 +4,7 @@
  * Return: int
  */
 
-int factorial(int n)
+float factorial(int n)
 {
 	if (n == 0)
 		return (1);
